requirements.txt
alembic==1.4.2
bcrypt==3.2.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.61.0
passlib==1.7.2
SQLAlchemy==1.3.18
pymysql==1.0.2
psycopg2==2.8.5
python-dotenv==0.19.0
python-jose==3.2.0
python-multipart==0.0.5
sentry-sdk==1.3.1
uvicorn==0.11.8